Refactored code based on the review comments received

+++ Week2_Assignment_Linking/dynamic/source/addition.cpp
@@ -1,5 +1,5 @@
 #include "add.h"
 
-int add(int a, int b) {
+double add(double a, double b) {
     return a + b;
 }
